[FEAT] label 레이아웃 추가 정리
더 큰 전용 라벨을 만들어서 적용하였다. 레이아웃도 여백 수정
