+++ client/src/components/issuelist/issueItem.js
@@ -13,7 +13,6 @@ import { COLOR_SUCCESS, COLOR_DANGER } from '../common/color.js';
 
 const IssueItemContainer = styled(ListItem)`
   display: flex;
-  padding: 8px 16px;
   align-items: flex-start;
 
   &:hover {
