+++ client/src/components/labellist/labelItemViewer.js
@@ -2,7 +2,7 @@ import React, { useContext } from 'react';
 import styled from 'styled-components';
 import { TOGGLE_ISEDIT } from '../../reducer/label';
 import { LabelContext } from '../common/context';
-import LabelItem from '../common/labelItem';
+import BigLabelItem from './biglabelitem';
 import LabelItemContainer from './itemContainer';
 
 const LabelItemViewer = (props) => {
@@ -10,7 +10,7 @@ const LabelItemViewer = (props) => {
 
   return (
     <LabelItemContainer>
-      <LabelItem label={props.label}/>
+      <BigLabelItem label={props.label}/>
       <p>{props.label.description}</p>
       <div>
         <button onClick={() => labelDispatch({type: TOGGLE_ISEDIT, payload: {id: props.label.id}})}>Edit</button>
